Refactored new Data Inflector into the already existing Data Processor.
