+++ tests/src/Unit/Plugin/RulesDataProcessor/TrimTest.php
@@ -2,16 +2,16 @@
 
 /**
  * @file
- * Contains \Drupal\Tests\rules\Unit\Plugin\DataInflector\TrimTest.
+ * Contains \Drupal\Tests\rules\Unit\Plugin\RulesDataProcessor\TrimTest.
  */
 
-namespace Drupal\Tests\rules\Unit\Plugin\DataInflector;
+namespace Drupal\Tests\rules\Unit\Plugin\RulesDataProcessor;
 
 use Drupal\Tests\rules\Unit\RulesUnitTestBase;
-use Drupal\rules\Plugin\DataInflector\Trim;
+use Drupal\rules\Plugin\RulesDataProcessor\Trim;
 
 /**
- * @coversDefaultClass \Drupal\rules\DataInflector\Trim
+ * @coversDefaultClass \Drupal\rules\RulesDataProcessor\Trim
  * @group rules
  */
 class TrimTest extends RulesUnitTestBase {
@@ -19,9 +19,9 @@ class TrimTest extends RulesUnitTestBase {
    * @dataProvider stringsProvider
    */
   public function testTrim($expectedMatchResult, $string) {
-    $inflector = new Trim([], 'rules_data_inflector_trim', []);
+    $processor = new Trim([], 'rules_data_processor_trim', []);
 
-    $this->assertSame($expectedMatchResult, $inflector->inflect($string));
+    $this->assertSame($expectedMatchResult, $processor->process($string));
   }
 
   public function stringsProvider() {
